+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog for PSTweetChat
 
+## 7 June 2019
+
++ Modified `New-PSTweetChatTranscript.ps1` to take CSV path as a parameter
+
+## May 2019
+
++ No PSTweetChat for this month.
+
 ## 1 March 2019
 
 + initial module files and organization